Set up project structure
This includes TypeScript, Webpack (for building the frame), ng-packagr (for building the library), Karma (for testing on Chrome), ESLint, Prettier, and GitHub Actions (for CI).

Also, switch to using the term "frame" instead of "binary"; the term "binary" to mean a compiled JavaScript file is a Bazelism not widely understood elsewhere.