Replace getAllAds with forEachInterestGroup (#107)
For efficiency, this works by streaming over the entries in the database. We take a synchronous callback to guarantee that it can't block, which would be a problem because the browser will automatically close the IndexedDB transaction in that case. The resulting API is slightly awkward, but workable and hard to misuse.

While we're in here migrating stuff to the cursor API, do the same to storeInterestGroup, to faciliate partial updates. (We'll need these for dailyUpdateUrl even if we don't allow them in the JavaScript API.) This does not result in a change in observable behavior for now.