Make database operations return boolean success (#108)
This is the first step of migrating the frame code (but not the library code) away from exceptions for circumstances that aren't bugs in the code (in particular, IndexedDB I/O errors). This will make the code more robust (since it's generally safe to assume that things won't throw) and facilitate error telemetry when we get to that point.